[arch][arm64] Implement lazy fpu/simd state loading
If the fpu/simd registers on this cpu were last used by the current
thread and the current thread last used the fpu/simd register on the
current cpu, then the load can be skipped.

We still always save the state, to avoid a situation where a cpu wants
to run a thread where the fpu/simd register state for that thread is
only available in the registers of another cpu.

Change-Id: I062aa0792180b9ec08a5dd95b5f4f6ba165167e1