Fill in missing PRI* macros for inttypes.h.
inttypes.h is supposed to define print macros for the various integers
defined in stdint.h. This change fills in the sized macros for all the
various suffixes. The 'o' suffixed variants are omitted because, at a
glance, lib/libc/printf.c does not support it.

Note to reviewers, these should be checked against the typedefs in
stdint.h and the prefixes defined for this libc's printf implementing.