[clang] Refactor declaration and use of extern linker variables
clang uses the C/C++ guarantee that &x != &y to make optimizations
in loops of the form for (xx = &x; xx != &y; xx++) where the first
check of the loop is skipped. This causes errors in cases where
x and y are defined in linker scripts and the two names alias
making &x == &y. Ideally we would resolve this by having a compiler
attribute to indicate that aliasing is possible for these variables
preventing clang from making this kind of optimization. However,
such attribute is not currently available in clang so in the
meantime we can avoid this optimization by declaring the variables
as unsized arrays and using them like pointers.

Change-Id: Ib7506676b8c98b22db0464c8e005e963fecdfaa1